requirements.txt
altgraph==0.17.4
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
cftime==1.6.3
chardet==4.0.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.0
cx_Freeze==7.0.0
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
executing==2.0.1
fonttools==4.50.0
idna==2.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
ipykernel==6.29.4
ipython==8.22.2
jedi==0.19.1
Jinja2==3.0.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
macholib==1.16.3
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
moviepy==1.0.3
nest-asyncio==1.6.0
netCDF4==1.6.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
Pillow==8.4.0
platformdirs==4.2.0
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
requests==2.25.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==69.2.0
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.2
threadpoolctl==3.4.0
tkdesigner==1.0.7
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
wheel==0.43.0
